+++ lib/glific/messages/message.ex
@@ -8,6 +8,7 @@ defmodule Glific.Messages.Message do
   alias Glific.{
     Contacts.Contact,
     Flows.Flow,
+    Groups.Group,
     Messages.MessageMedia,
     Partners.Organization,
     Tags.Tag,
@@ -36,6 +37,8 @@ defmodule Glific.Messages.Message do
           contact: Contact.t() | Ecto.Association.NotLoaded.t() | nil,
           user_id: non_neg_integer | nil,
           user: User.t() | Ecto.Association.NotLoaded.t() | nil,
+          group_id: non_neg_integer | nil,
+          group: Group.t() | Ecto.Association.NotLoaded.t() | nil,
           flow_id: non_neg_integer | nil,
           flow_object: Flow.t() | Ecto.Association.NotLoaded.t() | nil,
           media_id: non_neg_integer | nil,
@@ -71,6 +74,7 @@ defmodule Glific.Messages.Message do
     :bsp_message_id,
     :errors,
     :media_id,
+    :group_id,
     :send_at,
     :sent_at,
     :user_id,
@@ -107,6 +111,8 @@ defmodule Glific.Messages.Message do
     belongs_to :flow_object, Flow, foreign_key: :flow_id
     belongs_to :media, MessageMedia
     belongs_to :organization, Organization
+
+    belongs_to :group, Group
 
     many_to_many :tags, Tag, join_through: "messages_tags", on_replace: :delete