+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.7] - 2020-12-28
+### Added
+- Store messages sent to group in messages table
+- Add API to retrieve group conversations
+
 ## [0.8.6] - 2020-12-22
 ### Added
 - Stir usecase, computing score based on answers of survey